storage/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv

# Testing
pytest
httpx    # TestClient transport
